+++ packages/validator/src/.defaultsForValidator.js
@@ -21,11 +21,8 @@
 const defaults = {
   'shared': {
     'operations': {
-      'no_array_responses': 'error',
-      'parameter_order': 'warning',
     },
     'paths': {
-      'missing_path_parameter': 'error',
       'duplicate_path_parameter': 'warning',
       'snake_case_only': 'off',
       'paths_case_convention': ['error', 'lower_snake_case']
@@ -102,6 +99,9 @@ const deprecated = {
   'no_request_body_name': 'request-body-name (spectral rule)',
   'operation_id_case_convention': 'operation-id-case-convention (spectral rule)',
   'operation_id_naming_convention': 'operation-id-naming-convention (spectral rule)',
+  'no_array_responses': 'array-responses (spectral rule)',
+  'parameter_order': 'parameter-order (spectral rule)',
+  'missing_path_parameter': 'path-params and path-declarations-must-exist (spectral rule)',
 };
 
 const configOptions = {