+++ packages/ruleset/CHANGELOG.md
@@ -1,3 +1,31 @@
+# @ibm-cloud/openapi-ruleset [0.17.0](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset@0.16.0...@ibm-cloud/openapi-ruleset@0.17.0) (2022-04-19)
+
+
+### Features
+
+* **missing_path_parameter:** replace rule with spectral:oas rules ([#428](https://github.com/IBM/openapi-validator/issues/428)) ([4b7719f](https://github.com/IBM/openapi-validator/commit/4b7719f51489b80474e6ae4c8d5a56d139ec9f0b))
+
+# @ibm-cloud/openapi-ruleset [0.16.0](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset@0.15.0...@ibm-cloud/openapi-ruleset@0.16.0) (2022-04-19)
+
+
+### Features
+
+* **parameter-order:** add new 'parameter-order' rule ([#423](https://github.com/IBM/openapi-validator/issues/423)) ([3304225](https://github.com/IBM/openapi-validator/commit/33042257cab50dda1aa602cd3c3bd8ee33749b12))
+
+# @ibm-cloud/openapi-ruleset [0.15.0](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset@0.14.1...@ibm-cloud/openapi-ruleset@0.15.0) (2022-04-19)
+
+
+### Features
+
+* **array-responses:** add new 'array-responses' rule ([#422](https://github.com/IBM/openapi-validator/issues/422)) ([88cf48c](https://github.com/IBM/openapi-validator/commit/88cf48c1c9da04f49da6d82528b3c97be01ba356))
+
+## @ibm-cloud/openapi-ruleset [0.14.1](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset@0.14.0...@ibm-cloud/openapi-ruleset@0.14.1) (2022-04-19)
+
+
+### Bug Fixes
+
+* **content-entry-provided:** dont require content entry for 304 responses ([82c0c0a](https://github.com/IBM/openapi-validator/commit/82c0c0af2b2d41e8fc1ebdaf6ff2740a39d4ac06))
+
 # @ibm-cloud/openapi-ruleset [0.14.0](https://github.com/IBM/openapi-validator/compare/@ibm-cloud/openapi-ruleset@0.13.0...@ibm-cloud/openapi-ruleset@0.14.0) (2022-04-14)