+++ packages/validator/CHANGELOG.md
@@ -1,3 +1,58 @@
+# ibm-openapi-validator [0.68.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@0.67.0...ibm-openapi-validator@0.68.0) (2022-04-19)
+
+
+### Features
+
+* **missing_path_parameter:** replace rule with spectral:oas rules ([#428](https://github.com/IBM/openapi-validator/issues/428)) ([4b7719f](https://github.com/IBM/openapi-validator/commit/4b7719f51489b80474e6ae4c8d5a56d139ec9f0b))
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 0.17.0
+
+# ibm-openapi-validator [0.67.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@0.66.0...ibm-openapi-validator@0.67.0) (2022-04-19)
+
+
+### Features
+
+* **parameter-order:** add new 'parameter-order' rule ([#423](https://github.com/IBM/openapi-validator/issues/423)) ([3304225](https://github.com/IBM/openapi-validator/commit/33042257cab50dda1aa602cd3c3bd8ee33749b12))
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 0.16.0
+
+# ibm-openapi-validator [0.66.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@0.65.1...ibm-openapi-validator@0.66.0) (2022-04-19)
+
+
+### Features
+
+* **array-responses:** add new 'array-responses' rule ([#422](https://github.com/IBM/openapi-validator/issues/422)) ([88cf48c](https://github.com/IBM/openapi-validator/commit/88cf48c1c9da04f49da6d82528b3c97be01ba356))
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 0.15.0
+
+## ibm-openapi-validator [0.65.1](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@0.65.0...ibm-openapi-validator@0.65.1) (2022-04-19)
+
+
+
+
+
+### Dependencies
+
+* **@ibm-cloud/openapi-ruleset:** upgraded to 0.14.1
+
 # ibm-openapi-validator [0.65.0](https://github.com/IBM/openapi-validator/compare/ibm-openapi-validator@0.64.0...ibm-openapi-validator@0.65.0) (2022-04-14)